Extract CheckLicense and SubmitLicense API tests and CheckLicense App tests
